finish the permission module
