+++ src/main/java/com/google/personalhealthrecordingplateform/configs/RedisConfig.java
@@ -1,0 +1,2 @@
+package com.google.personalhealthrecordingplateform.configs;public class RedisConfig {
+}
