--- src/main/java/com/google/personalhealthrecordingplateform/configs/security/RedisConfig.java
@@ -1,23 +1,0 @@
-package com.google.personalhealthrecordingplateform.configs.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-
-@Configuration
-public class RedisConfig {
-
-    @Bean
-    public RedisTemplate getRedisTemplate() {
-        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-        return redisTemplate;
-    }
-
-    @Bean
-    public LettuceConnectionFactory lettuceConnectionFactory() {
-        LettuceConnectionFactory factory = new LettuceConnectionFactory();
-        factory.setPassword("123456");
-        return factory;
-    }
-}
