+++ src/main/java/com/google/personalhealthrecordingplateform/controllers/SysPermissionController.java
@@ -1,0 +1,2 @@
+package com.google.personalhealthrecordingplateform.controllers;public class SysPermissionController {
+}
