+++ src/main/java/com/google/personalhealthrecordingplateform/entity/SysPermission.java
@@ -2,9 +2,13 @@ package com.google.personalhealthrecordingplateform.model;
 
 import lombok.Data;
 
+import java.io.Serializable;
+
 @Data
-public class SysRole {
+public class SysPermission implements Serializable {
     private Long id;
+
     private String label;
+
     private String code;
 }
