+++ src/main/java/com/google/personalhealthrecordingplateform/entity/SysRole.java
@@ -2,11 +2,11 @@ package com.google.personalhealthrecordingplateform.model;
 
 import lombok.Data;
 
+import java.io.Serializable;
+
 @Data
-public class SysPermission {
+public class SysRole implements Serializable {
     private Long id;
-
     private String label;
-
     private String code;
 }
