+++ src/main/java/com/google/personalhealthrecordingplateform/mapper/SysPermissionMapper.java
@@ -1,0 +1,2 @@
+package com.google.personalhealthrecordingplateform.mapper;public class SysPermissionMapper {
+}
