+++ src/main/java/com/google/personalhealthrecordingplateform/security/SecurityConfig.java
@@ -48,11 +48,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http.addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
     }
 
-//    @Override
-//    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-//        auth.userDetailsService(userDetailsService())
-//                .passwordEncoder(passwordEncoder());
-//    }
+
 
     @Bean
     public PasswordEncoder passwordEncoder() {
