+++ src/main/java/com/google/personalhealthrecordingplateform/security/content/SecurityContent.java
