+++ src/main/java/com/google/personalhealthrecordingplateform/security/handler/JwtAuthenticationEntryPoint.java
