+++ src/main/java/com/google/personalhealthrecordingplateform/services/SysPermissionService.java
@@ -1,0 +1,2 @@
+package com.google.personalhealthrecordingplateform.services;public class SysPermissionService {
+}
