+++ src/main/java/com/google/personalhealthrecordingplateform/services/impl/SysPermissionServiceImpl.java
@@ -1,0 +1,2 @@
+package com.google.personalhealthrecordingplateform.services.impl;public class SysPermissionServiceImpl {
+}
